+++ packages/cli/src/command-registry.ts
@@ -1,0 +1,36 @@
+import type { CommandRegistryEntry } from "./types.ts";
+
+export const commandRegistry = [
+  { kind: "init", primary: "harness init", resultEnvelope: "CliResult/v1" },
+  { kind: "new-task", primary: "harness new-task --title <title> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "status-set", primary: "harness task status set <id> <status>", resultEnvelope: "CliResult/v1" },
+  { kind: "progress-append", primary: "harness task progress append <id> --text <text>", resultEnvelope: "CliResult/v1" },
+  { kind: "task-archive", primary: "harness task archive <id> --reason <reason>", resultEnvelope: "CliResult/v1" },
+  { kind: "task-supersede", primary: "harness task supersede <old-id> --title <title> [--slug <slug>]", resultEnvelope: "CliResult/v1" },
+  { kind: "task-delete", primary: "harness task delete (--soft|--hard) <id> --reason <reason>", resultEnvelope: "CliResult/v1" },
+  { kind: "task-reopen", primary: "harness task reopen <id> --reason <reason>", resultEnvelope: "CliResult/v1" },
+  { kind: "task-review", primary: "harness task-review <id> [--reviewer <id>]", resultEnvelope: "CliResult/v1" },
+  { kind: "task-complete", primary: "harness task-complete <id> --ci passed|failed", resultEnvelope: "CliResult/v1" },
+  { kind: "task-list", primary: "harness task list [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "status", primary: "harness status --json", resultEnvelope: "CliResult/v1" },
+  { kind: "check", primary: "harness check [--profile source-package|private-harness|target-project] [--strict] [--post-merge] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "governance-rebuild", primary: "harness governance rebuild [--dry-run|--archive|--apply] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "lesson-promote", primary: "harness lesson-promote <task-id> <candidate-id> [--dry-run|--apply] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "lesson-sediment", primary: "harness lesson-sediment <task-id> <candidate-id> [--dry-run] [--title <title>] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-list", primary: "harness preset list [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-inspect", primary: "harness preset inspect <id> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-check", primary: "harness preset check <id> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-install", primary: "harness preset install <folder> [--project] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-seed", primary: "harness preset seed [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-audit", primary: "harness preset audit [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-uninstall", primary: "harness preset uninstall <id> [--project] [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-run", primary: "harness preset run <id> <plan|scaffold|check> --task <id> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "preset-action", primary: "harness preset action <id> <action> --task <id> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-list", primary: "harness module list [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-inspect", primary: "harness module inspect <key> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-register", primary: "harness module register <key> --title <title> --scope <path> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-scaffold", primary: "harness module scaffold <key> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-unregister", primary: "harness module unregister <key> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "module-step", primary: "harness module-step <key> <step> --state <state> [--json]", resultEnvelope: "CliResult/v1" },
+  { kind: "gui", primary: "harness gui", resultEnvelope: "CliResult/v1" }
+] as const satisfies ReadonlyArray<CommandRegistryEntry>;
